+++ packages/cli/src/lib/upload/upload-command.ts
@@ -1,11 +1,9 @@
-import ansis from 'ansis';
 import type { ArgumentsCamelCase, CommandModule } from 'yargs';
 import { type UploadOptions, upload } from '@code-pushup/core';
 import { logger } from '@code-pushup/utils';
-import { CLI_NAME } from '../constants.js';
 import {
-  renderIntegratePortalHint,
-  uploadSuccessfulLog,
+  printCliCommand,
+  renderPortalHint,
 } from '../implementation/logging.js';
 
 export function yargsUploadCommandObject() {
@@ -14,18 +12,16 @@ export function yargsUploadCommandObject() {
     command,
     describe: 'Upload report results to the portal',
     handler: async <T>(args: ArgumentsCamelCase<T>) => {
-      logger.info(ansis.bold(CLI_NAME));
-      logger.debug(`Running ${ansis.bold(command)} command`);
+      printCliCommand(command);
 
       const options = args as unknown as UploadOptions;
       if (options.upload == null) {
-        renderIntegratePortalHint();
-        throw new Error('Upload configuration not set');
+        logger.newline();
+        renderPortalHint();
+        logger.newline();
+        throw new Error('Upload to Portal is missing configuration');
       }
-      const report = await upload(options);
-      if (report?.url) {
-        uploadSuccessfulLog(report.url);
-      }
+      await upload(options);
     },
   } satisfies CommandModule;
 }